requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
geoalchemy2==0.14.3
python-multipart==0.0.6
redis==4.6.0
fastapi-cache2[redis]==0.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings>=2.7.0
email-validator>=2.1.0
bcrypt==4.0.1